DAO USER

+++ src/main/java/myPackage/persistence/DAO/UserDAO.java
@@ -1,10 +1,11 @@
 package myPackage.persistence.DAO;
 
 import myPackage.persistence.Entity.UserEntity;
+import org.springframework.data.mongodb.repository.MongoRepository;
 
 /**
  * Created by thebous on 03/07/17.
  */
-public interface UserDAO extends MongoRepository<UserEntity, long>{
+public interface UserDAO extends MongoRepository<UserEntity, Integer> {
 
 }
